seperate initilize neo4j, run it before all other tasks
